+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWTVR58GMARGV79KBEAK8MG5"
+        "01KWV652VXCDXE6A65CDV0A48R"
     )
     r"""The unique identifier of the data source"""