+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWTVR58MQ3WN60HTPSF8YNWM"
+        "01KWV652W0JAPVR32R18JG5YT1"
     )
     r"""The unique identifier of the data source"""